in cjumpchain, changed current_state_for_uncond_probs[0] = q_t - number of ancestral species in boreal in sample history at time t current_state_for_uncond_probs[1] = r_t - number of ancestral species in neotropics in sameple history at time t
to 

current_state_for_uncond_probs[0] = q_t
current_state_for_uncond_probs[1] = r_t                   

Added back the use of CombOfNegative in CalculateLogPi

There is an error on page 20. Equations 26 and 27 refer to the same set of equations

Outlined:
def GetInfoForForwardProbMIG(A,level). This function will get the correct input values for  ForwardProbMIG. As of now,
it outputs one value for testing purposes, and I just input GetInfoForProbKMigrationInL(-1,-1) to signify that this function is 
not yet implemented. In order for this function to work, it needs 
SampleFromIS to be working. Since Perry is finishing cjumpchain, I am leaving that up to him unless he informs me otherwise.

Wrote:
ForwardRateSTT
ForwardRateSBB
ForwardRateSBT
ForwardRateSBArrowT
See Equations 26 and 27 in the paper for the descriptions of these equations
ProbKMigrationInL(level,l,A,sigma):     Returns the forward probability of the k migrations that occur in the level level. Defined 					       
in equation 29. For now input level= -1, A= -1, until SampleFromIS is implemented
PhiJK: a helper function to ProbKMigrationInL. I need confirmation from Ganesh that I implemented the cases when j=1 or j=k correctly
Phi: Phi(i) is the forward rate of SBarrowT right before the i-th migration
ForwardProbMIG(A,sigma): Calculates the probability of all SBArrowT events which is equivalent to P(MIG(theta)|sigma). As of now
		     input A=-1 until SampleFromIS is implemented
    

Sometimes there is a precision error and ProbKMigrationInL returns 0 when it should return a very small number, however this is
unlikely to be a problem since then the sampling weight will be zero (no divisions by zero)