Changed ForwardProbMIG name to ForwardProbA. Now ForwardProbMIG calculates eq. 30.
in cjumpchain one of the comments for SampleFromIS is wrong.

 For the next level, (i.e., level 3) the initial character state
        assignment will be: {1:0, 4:1, 5:1} (since 2 & 3 were in state
        1, their parent 5 also will be in state 1).

{1:0, 4:1, 5:1} as the initial character state assignment

 I changed this to

 For the next level, (i.e., level 3) the initial character state
        assignment will be: {1:0, 4:1, 5:1} (since 2 & 3 were in state
        0, their parent 5 also will be in state 0).

{1:0, 4:1, 5:0} as the initial character state assignment

Re-wrote def GetInfoForForwardProbMIG(A,level): into a multiple functions compatible with A's tree structure:
	GetNumberOfLevels(A):
	GetBranchLength(A,level):
	GetNumberOfMigrationEvents(A,level):
	GetQ_bbAndQ_bt(A,level):
		I need help with this one

Different inputs, but same function:
	ProbKMigrationInL(level,l,k,q_bb,q_bt,sigma)
	PhiJK(j,k,(r_t,q_bb,q_b_arrow_t,q_bt),sigma)
	Phi(i,(r_t,q_bb,q_b_arrow_t,q_bt),sigma):

Implemented GetH(X,i,A,sigma):  Gives the rate of event X in the ith level
Implemented GetHBar(i,A,sigma): gives the sum of the rates of all possible speciation events in the level i
Implemented GetYi(A,i): returns the speciation event that ends level i

Theoretically, ForwardProbA should be working now, (which may mean all of is_likelihood is implemented and working)

Added to MakeTransitionMatrixForLevel documentation

Didn't like PickNextStateofChain implementation. May work, but not intuitive to me, so re-wrote.

I think there is an error in the last lines of code of ChooseLineageAndUpdateDelta. If lineages are already in state 1, they shouldn't be included in the random_list.

As code was:

for i in range(len(current_delta)):
            if (i != next_coalescing_lineages[0].index and i != next_coalescing_lineages[1].index):
                random_list.append(i)

random.shuffle(random_list)

        lineage_to_migrate = random_list[0]
        if (current_delta[lineage_to_migrate] == 1): 
            #the if statement is really redundant
            current_delta[lineage_to_migrate] = 0
        return random_list[0]

So I changed the line to 
	            if (i != next_coalescing_lineages[0].index and i != next_coalescing_lineages[1].index and current_delta[i]!=1):

I also added an assert statement to ChooseLineageAndUpdateDelta

changed delta structure to reflect the fact that delta is appended to event history and should be consistent with the event_history description in is_classes. So I changed current_delta from " a list of of character states (boreal or tropical)"
to " a map of lineage indices to character states". Made modifications in the code to reflect the fact that delta is a map, not a list.

added all_delta, a map of all lineages at all levels to their states. As it is updated as A is created, it will eventually be a map of all lineages to their states right after they came into existence

MoveToNextLevel finished

Edited PrepareTree

in SampleFromIS

            if(index_of_next_state==len(transition_matrix_for_the_level):
               whether_in_the_same_level == False
            #whether_in_the_same_level = WhetherInTheSameLevel(state_of_cond_jump_chain,next_state_of_cond_jump_chain)

Since next_state_of_cond_jump_chain = index_in_transition_matrix_to_state[len(index_in_transition_matrix_to_state)] is undefined. A kappa event has an index of len(index_in_transition_matrix_to_state) which is outside of the range of index_in_transition_matrix_to_state.

created a Test function to get the inputs for SampleFromIS

SampleFromIS when called from the Test function in cjumpchain is returning a value and is filling out the event_histories for each level

I haven't gotten around to debugging is_likelihood yet.