Fixed parenthesis error in CalculateLogPi, and removed an erroneous parameter
There is an error in is_likelihood because levels[num_leaves-level] is the actual level, not levels[level]
Also, levels start at 2, not 1

GetFactorialOfNegative needed to be modified. That was causing the hang-up. It now has three if statements for if the 
input (after adding 1) is negative, 0, or positive. Also, the number of terms summed had to be lowered to 20 at the cost of 
precision to aid in faster calculations

Changed the assert Statement in SampleFromIs to 
assert(1.0-1e5<(sum(transition_matrix_for_the_level[index_of_current_state]))< 1.0+1e-5)

with sigma=(0.10000000000000001, 0.010000000000000012, 50, 0.040000000000000003, 0.10000000000000001)
no migrations

with sigma=(0.10000000000000001, 0.080000000000000012, 50, 0.090000000000000003, 0.10000000000000001)
all migrations in level 4

with sigma=(0.10000000000000001, 0.010000000000000012, 50, 0.090000000000000003, 0.10000000000000001)
all migrations in level 2

with sigma=(0.10000000000000001, 0.040000000000000012, 50, 0.090000000000000003, 0.10000000000000001)
migrations in levels 0 and 1

I will use this last sigma and G for testing is_likelihood

I need to see if all_delta is working as I want it to