is_likelihood is running without errors but is not finished.
How can hbar be calculated for the most recent level if q_bb, q_bt, qbb are not known?
for now, I am leaving out the part of the code of hbar(n) and just making that factor 1 in the multiplication

How to calculate the probability of zero migrations in a level? If there are no migrations in the level, for now I am having the code say the probability of zero migrations in that level is 1.

Even though exact times aren't assigned to migration events, they do occur in certain levels, so in that sense, are times assigned?