Removing redundancies in ApplyEvent added cases into IsValidState
I don't know what this part of the cjumpchain code under MakeTransitionMatrixForLevel was doing, and it was giving me 
error statements, so commented it out. Please explain this part of the code to me.

        #Testing if sum of rows = P(F|z), really normalizing against P(F|z)
        #testing = transition_matrix[row].copy()
        #testing2 = transition_matrix[row].copy()
        #Equal is defined in this class to test two numbers are so close
        #they're considered "equal" - use this is b/c == is hard
        #to apply to float numbers. 
        #assert(Equal(testing.sum(), solution[row]))
        #for j in range(len(testing)):
        #    testing[j] = testing[j]/solution[row]
        #Normalize(testing2)  
       #assert(allclose(testing,testing2))

Checked up to and including ApplyEvent

PickNextStateOfChain is messed up. I will have to look at this later